+++ editor/webview/views/project/index.scss
@@ -167,165 +167,4 @@ $transition: 0.3s all;
             }
         }
     }
-
-    .git-btn {
-        position: relative;
-        display: grid;
-        grid-template-areas:
-            "icon branch"
-            "icon commit";
-        text-align: left;
-        grid-template-columns: 32px auto;
-        @media screen and (max-width: $break-sm) {
-            grid-template-columns: 24px auto;
-        }
-        svg {
-            grid-area: icon;
-            margin: 0 !important;
-        }
-        > .branch {
-            grid-area: branch;
-            margin-top: -4px;
-            @media screen and (max-width: $break-sm) {
-                display: none;
-            }
-        }
-        > .commit {
-            grid-area: commit;
-            font-size: $font-size-s;
-            margin-top: -2px;
-            margin-bottom: -3px;
-            opacity: 0.8;
-            @media screen and (max-width: $break-sm) {
-                display: none;
-            }
-        }
-
-        .push,
-        .pull,
-        .alert {
-            position: absolute;
-            z-index: 1;
-            left: 30px;
-            height: 20px;
-            top: 5px;
-            svg {
-                height: 100%;
-                width: 100%;
-            }
-        }
-
-        .pull {
-            transform: rotate(45deg);
-            color: $green;
-        }
-
-        .push {
-            transform: rotate(225deg);
-            color: $red;
-        }
-
-        .alert {
-            background-color: colorOpacity($red, 0.2);
-            border-radius: 50%;
-            color: $red;
-            left: 32px;
-            top: 3px;
-            height: 16px;
-            width: 16px;
-        }
-    }
-
-    .git-dialog {
-        header {
-            display: grid;
-            grid-template-columns: 45px auto;
-            grid-template-areas:
-                "icon url"
-                "icon branch"
-                "icon commit";
-
-            > svg {
-                grid-area: icon;
-                align-self: center;
-                height: 32px;
-                width: 32px;
-            }
-            > a {
-                grid-area: url;
-                color: colorOpacity(white, 0.7);
-                font-size: $font-size-s;
-                overflow: hidden;
-                text-overflow: ellipsis;
-                &:hover {
-                    color: colorOpacity(white, 1);
-                }
-            }
-            > div:nth-child(3) {
-                grid-area: branch;
-            }
-            > div:last-child {
-                grid-area: commit;
-                font-size: $font-size-s;
-                overflow: hidden;
-                text-overflow: ellipsis;
-            }
-        }
-        > .author {
-            display: flex;
-            font-size: $font-size-s;
-            grid-template-columns: 45px auto 45px;
-            background-color: colorOpacity(white, 0.1);
-            margin: spacing() 0;
-            padding: spacing();
-            border-radius: $border-radius;
-            border: $border-width solid colorOpacity(white, 0.3);
-            > svg {
-                align-self: center;
-                justify-self: center;
-                margin-right: spacing();
-            }
-            > *:nth-child(2) {
-                flex: 1;
-                display: flex;
-                flex-direction: column;
-                overflow: hidden;
-                text-overflow: ellipsis;
-                > input {
-                    margin-bottom: spacing();
-                }
-            }
-            form {
-                > p {
-                    svg {
-                        color: $red;
-                        vertical-align: middle;
-                        height: 18px;
-                        width: 18px;
-                    }
-                    margin-top: 0;
-                    display: flex;
-                    align-items: center;
-                    gap: spacing(0.5);
-                }
-                > div:last-child {
-                    display: flex;
-                    flex-direction: row-reverse;
-                    > * {
-                        flex: 1;
-                    }
-                }
-            }
-        }
-
-        > .changes {
-            display: flex;
-            flex-direction: column;
-            padding: spacing() 0;
-        }
-        > div:last-child {
-            display: flex;
-            justify-content: space-between;
-        }
-    }
 }
