+++ editor/webview/views/project/index.ts
@@ -3,12 +3,12 @@ import "./index.css";
 import { Editor } from "../editor";
 import { FileTree } from "../file-tree";
 import { Console } from "../console";
+import { DELETE_ALL_PACKAGES_ID, RUN_PROJECT_ID } from "../../../constants";
+import GitWidget from "./git-widget";
 
 import type { Project as TypeProject } from "../../../api/projects/types";
 import type typeRPC from "../../../../src/webview";
 import type api from "../../../api";
-import { DELETE_ALL_PACKAGES_ID, RUN_PROJECT_ID } from "../../../constants";
-import git from "../../../api/git";
 
 declare var rpc: typeof typeRPC<typeof api>;
 
@@ -28,13 +28,13 @@ export class Project {
     };
     console = new Console();
 
+    private gitWidget = new GitWidget();
+
     private tabsContainer = document.createElement("ul");
     private editorsContainer = document.createElement("div");
 
     private currentFile: string;
     private editors: Editor[] = [];
-
-    private gitBtn: HTMLButtonElement;
 
     constructor() {
         this.fileTree.instance.onItemSelect = (item) => {
@@ -185,6 +185,8 @@ export class Project {
         if (project === this.project) return;
 
         this.project = project;
+        this.gitWidget.project = project;
+
         this.fileTree.instance.setBaseDirectory(project.location);
 
         this.editors = [];
@@ -294,268 +296,6 @@ export class Project {
         return container;
     }
 
-    private async renderGitButton(pull = false) {
-        this.gitBtn = document.createElement("button");
-        this.gitBtn.classList.add("text", "text-and-icon", "git-btn");
-
-        const [gitIcon, arrowIcon] = await Promise.all([
-            (await fetch("assets/icons/git.svg")).text(),
-            (await fetch("assets/icons/arrow.svg")).text()
-        ]);
-        this.gitBtn.innerHTML = gitIcon;
-
-        const [branch, commit] = await Promise.all([
-            rpc().git.currentBranch(this.project),
-            rpc().git.log(this.project, 1)
-        ]);
-        const currentBranch = branch || "DETACHED";
-        const currentCommit = commit?.at(0)?.oid || "";
-
-        this.gitBtn.addEventListener("click", () =>
-            this.renderGitDialog(gitIcon, currentBranch, currentCommit)
-        );
-
-        const branchContainer = document.createElement("div");
-        branchContainer.classList.add("branch");
-        branchContainer.innerText = currentBranch;
-        this.gitBtn.append(branchContainer);
-
-        const commitContainer = document.createElement("div");
-        commitContainer.classList.add("commit");
-        commitContainer.innerText = currentCommit.slice(0, 7);
-        this.gitBtn.append(commitContainer);
-
-        if (pull) {
-            const pullIcon = document.createElement("div");
-            pullIcon.classList.add("pull");
-            pullIcon.innerHTML = arrowIcon;
-            this.gitBtn.prepend(pullIcon);
-
-            setTimeout(async () => {
-                const maybeError = await rpc().git.pull(this.project);
-                pullIcon.remove();
-                if (maybeError && maybeError?.error) {
-                    const alertIcon = document.createElement("div");
-                    alertIcon.classList.add("alert");
-                    alertIcon.innerHTML = await (
-                        await fetch("assets/icons/alert.svg")
-                    ).text();
-                    this.gitBtn.prepend(alertIcon);
-                } else {
-                    this.gitBtn.replaceWith(await this.renderGitButton());
-                }
-            }, 500);
-        }
-
-        return this.gitBtn;
-    }
-
-    private async renderGitDialog(
-        icon: string,
-        branch: string,
-        commit: string
-    ) {
-        const dialog = document.createElement("div");
-        dialog.classList.add("dialog");
-
-        const container = document.createElement("div");
-        container.classList.add("git-dialog");
-
-        const gitInfo = document.createElement("header");
-        const remote = this.project.gitRepository.url;
-        gitInfo.innerHTML = `
-            ${icon}
-            <a href="${remote}" target="_blank">${remote.slice(0, -".git".length)}</a>
-            <div>${branch}</div>
-            <div>${commit}</div>`;
-
-        container.append(gitInfo);
-
-        const authorContainer = document.createElement("div");
-        authorContainer.classList.add("author");
-        container.append(authorContainer);
-
-        const [userIcon, editIcon, closeIcon, checkIcon] = await Promise.all([
-            (await fetch("assets/icons/user.svg")).text(),
-            (await fetch("assets/icons/edit.svg")).text(),
-            (await fetch("assets/icons/close.svg")).text(),
-            (await fetch("assets/icons/check.svg")).text()
-        ]);
-
-        const renderAuthorInputs = async () => {
-            authorContainer.innerHTML = `
-                ${userIcon}`;
-
-            const form = document.createElement("form");
-
-            if (!this.project.gitRepository.name) {
-                const alert = document.createElement("p");
-                alert.innerHTML = `${await (await fetch("assets/icons/alert.svg")).text()}
-                    No git user.name`;
-                form.append(alert);
-            }
-
-            const nameLabel = document.createElement("label");
-            nameLabel.innerText = "Name";
-            form.append(nameLabel);
-
-            const nameInput = document.createElement("input");
-            nameInput.value = this.project.gitRepository.name || "";
-            form.append(nameInput);
-
-            const emailInputLabel = document.createElement("label");
-            emailInputLabel.innerText = "Email";
-            form.append(emailInputLabel);
-
-            const emailInput = document.createElement("input");
-            emailInput.value = this.project.gitRepository.email || "";
-            emailInput.type = "email";
-            form.append(emailInput);
-
-            const buttonGroup = document.createElement("div");
-
-            const confirmButton = document.createElement("button");
-            confirmButton.classList.add("text");
-            confirmButton.addEventListener("click", async () => {
-                this.project.gitRepository.name = nameInput.value;
-                this.project.gitRepository.email = emailInput.value;
-                renderAuthorInfo();
-                await rpc().projects.update(this.project);
-                this.gitBtn.replaceWith(await this.renderGitButton());
-            });
-            confirmButton.innerHTML = checkIcon;
-            buttonGroup.append(confirmButton);
-
-            if (this.project.gitRepository.name) {
-                const cancelButton = document.createElement("button");
-                cancelButton.classList.add("text", "danger");
-                cancelButton.addEventListener("click", () => {
-                    renderAuthorInfo();
-                });
-                cancelButton.innerHTML = closeIcon;
-                buttonGroup.append(cancelButton);
-            } else {
-                const filler = document.createElement("div");
-                buttonGroup.append(filler);
-            }
-
-            form.append(buttonGroup);
-
-            authorContainer.append(form);
-        };
-
-        const renderAuthorInfo = async () => {
-            authorContainer.innerHTML = `
-                ${userIcon}
-                <div>
-                    <div>${this.project.gitRepository.name || "<b>No username</b>"}</div>
-                    <div>${this.project.gitRepository.email || "<b>No email</b>"}</div>
-                </div>`;
-
-            const editButton = document.createElement("button");
-            editButton.addEventListener("click", () => renderAuthorInputs());
-            editButton.classList.add("small", "text");
-            editButton.innerHTML = editIcon;
-            authorContainer.append(editButton);
-        };
-
-        if (
-            this.project.gitRepository.name ||
-            this.project.gitRepository.email
-        ) {
-            await renderAuthorInfo();
-        } else {
-            await renderAuthorInputs();
-        }
-
-        const confirmButton = document.createElement("button");
-        const changesContainer = document.createElement("div");
-        changesContainer.classList.add("changes");
-        changesContainer.innerText = "Calculating diffs...";
-        container.append(changesContainer);
-
-        const getChanges = async () => {
-            const changes = await rpc().git.changes(this.project);
-            changesContainer.innerText = "";
-
-            const hasChanges = Object.values(changes).some(
-                (arr) => arr.length !== 0
-            );
-
-            if (!hasChanges) {
-                changesContainer.innerHTML = `<h3>No changes</h3>`;
-            } else {
-                Object.entries(changes).forEach(([status, files]) => {
-                    if (files.length === 0) return;
-
-                    const subtitle = document.createElement("h3");
-                    subtitle.innerText =
-                        status.at(0).toUpperCase() + status.slice(1);
-                    changesContainer.append(subtitle);
-
-                    const ul = document.createElement("ul");
-                    const lis = files.map((file) => {
-                        const li = document.createElement("li");
-                        li.innerText = file;
-                        return li;
-                    });
-                    ul.append(...lis);
-                    changesContainer.append(ul);
-                });
-
-                const commitMessageInputLabel = document.createElement("label");
-                commitMessageInputLabel.innerText = "Commit Message";
-                changesContainer.append(commitMessageInputLabel);
-
-                const commitMessageInput = document.createElement("input");
-                changesContainer.append(commitMessageInput);
-
-                confirmButton.disabled = false;
-                confirmButton.addEventListener("click", async () => {
-                    dialog.remove();
-
-                    const arrowIcon = await (
-                        await fetch("assets/icons/arrow.svg")
-                    ).text();
-
-                    let pushIcon = document.createElement("div");
-                    pushIcon.classList.add("push");
-                    pushIcon.innerHTML = arrowIcon;
-                    this.gitBtn.prepend(pushIcon);
-
-                    setTimeout(async () => {
-                        await rpc().git.push(
-                            this.project,
-                            commitMessageInput.value
-                        );
-                        this.gitBtn.replaceWith(await this.renderGitButton());
-                    }, 500);
-                });
-            }
-        };
-
-        const buttonGroup = document.createElement("div");
-
-        const cancelButton = document.createElement("button");
-        cancelButton.classList.add("text");
-        cancelButton.innerText = "Close";
-        cancelButton.addEventListener("click", () => {
-            dialog.remove();
-        });
-        buttonGroup.append(cancelButton);
-
-        confirmButton.disabled = true;
-        confirmButton.innerText = "Push";
-        buttonGroup.append(confirmButton);
-
-        container.append(buttonGroup);
-
-        dialog.append(container);
-        this.container.append(dialog);
-
-        setTimeout(getChanges, 500);
-    }
-
     private async renderToolbar() {
         const container = document.createElement("div");
         container.classList.add("top-bar");
@@ -587,9 +327,7 @@ export class Project {
         projectTitle.innerText = this.project.title;
         leftSide.append(projectTitle);
 
-        if (this.project.gitRepository) {
-            leftSide.append(await this.renderGitButton(true));
-        }
+        leftSide.append(await this.gitWidget.renderButton(true));
 
         container.append(leftSide);
 
@@ -665,6 +403,7 @@ export class Project {
 
     async render() {
         this.container = document.createElement("div");
+        this.gitWidget.parentContainer = this.container;
         this.container.classList.add("project");
 
         this.container.append(await this.renderToolbar());
